+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Write the date in place of the "Unreleased" in the case a new version is release
 ### Fixed
 
 - Column names in `TableStructure` are explicitly converted to strings.
+- Ensure that structural dtype arrays read with `CSVAdapter` have two dimensions, `(n, 1)`.
 
 
 ## v0.1.6 (2025-09-29)